Consolidate proto and protocols into mod protocol
