+++ codegen-server/src/main/kotlin/software/amazon/smithy/rust/codegen/server/smithy/ServerRuntimeType.kt
@@ -18,7 +18,7 @@ object ServerRuntimeType {
         ServerCargoDependency.smithyHttpServer(runtimeConfig).toType().resolve("routing::Router")
 
     fun protocol(name: String, path: String, runtimeConfig: RuntimeConfig) =
-        ServerCargoDependency.smithyHttpServer(runtimeConfig).toType().resolve("proto::$path::$name")
+        ServerCargoDependency.smithyHttpServer(runtimeConfig).toType().resolve("protocol::$path::$name")
 
     fun protocol(runtimeConfig: RuntimeConfig) = protocol("Protocol", "", runtimeConfig)
 }
