+++ codegen-server/src/main/kotlin/software/amazon/smithy/rust/codegen/server/smithy/generators/ServerRuntimeTypesReExportsGenerator.kt
@@ -50,7 +50,7 @@ class ServerRuntimeTypesReExportsGenerator(
             }
 
             pub use #{SmithyHttpServer}::instrumentation;
-            pub use #{SmithyHttpServer}::proto;
+            pub use #{SmithyHttpServer}::protocol;
 
             pub use #{SmithyHttpServer}::Extension;
             """,
