+++ rust-runtime/aws-smithy-http-server/src/protocol/aws_json/runtime_error.rs
@@ -3,10 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use crate::proto::aws_json_11::AwsJson1_1;
+use crate::protocol::aws_json_11::AwsJson1_1;
 use crate::response::IntoResponse;
 use crate::runtime_error::{InternalFailureException, INVALID_HTTP_RESPONSE_FOR_RUNTIME_ERROR_PANIC_MESSAGE};
-use crate::{extension::RuntimeErrorExtension, proto::aws_json_10::AwsJson1_0};
+use crate::{extension::RuntimeErrorExtension, protocol::aws_json_10::AwsJson1_0};
 use http::StatusCode;
 
 use super::rejection::{RequestRejection, ResponseRejection};
