+++ rust-runtime/aws-smithy-http-server/src/protocol/rest/mod.rs
