+++ rust-runtime/aws-smithy-http-server/src/protocol/rest_xml/mod.rs
