+++ rust-runtime/aws-smithy-http-server/src/protocol/rest_xml/runtime_error.rs
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use crate::proto::rest_xml::RestXml;
+use crate::protocol::rest_xml::RestXml;
 use crate::response::IntoResponse;
 use crate::runtime_error::InternalFailureException;
 use crate::{extension::RuntimeErrorExtension, runtime_error::INVALID_HTTP_RESPONSE_FOR_RUNTIME_ERROR_PANIC_MESSAGE};
